Add "the" before Micronaut Framework
In previous guides I've proofread, <h2> level headings were in title case and <h3> level headings were in sentence case. If that is a style convention you wish to keep consistent, this <h2> would have "framework" capitalized. I also added "the" before "Micronaut." It also just occurred to me as I was typing this note that all headings in the guides are transformed to uppercase text via the CSS, so it actually doesn't matter how headings are capitalized in the Markdown. *facepalm