%s/Micronaut Framework/the Micronaut framework/ (#538)
* %s/Micronaut Framework/the Micronaut framework/

* Capitalization update

In this case, "framework" is part of a title, so it remains capitalized.

* Capitalization update

Framework is part of a title

* Capitalization

Framework is part of a title in this instance

* Added hyphens

Step-by-step is hyphenated when used as an adjective

* Micronaut already used as an adjective

In this case, Micronaut modifies "caching annotations," so it is being used correctly as an adjective. The only time you need to add "framework" is in cases where Micronaut is being used as a noun.

* Capitalization update

"Error" is not a proper noun, so it does not need to be capitalized in a sentence.

* Capitalization and punctuation update

Framework is correctly capitalized because this is a title; "Event-Driven Applications" is also part of a title, so it needs to be capitalized as well. Also because "event-driven" is being used as a compound adjective, it needs to be hyphenated.

* Add "the" before Micronaut Framework

In previous guides I've proofread, <h2> level headings were in title case and <h3> level headings were in sentence case. If that is a style convention you wish to keep consistent, this <h2> would have "framework" capitalized. I also added "the" before "Micronaut." It also just occurred to me as I was typing this note that all headings in the guides are transformed to uppercase text via the CSS, so it actually doesn't matter how headings are capitalized in the Markdown. *facepalm

* Capitalization update

Framework is capitalized when used in the title

* Title in title case

Capitalized the words that needed capitalizing in the title

* Title case update

Capitalized the words in the title that needed to be capitalized

* Capitalization update

Put guide title in title case

* Capitalization update

Put guide title in title case

* Capitalization update

Multitenancy and microservices are not proper nouns, so I put them in lowercase.

* Title case

Put the guide title in title case

* Capitalization update

title case

* Capitalization update

title case

* Capitalization update

Title case for <h2>

* Proofreading (#550)

Minor changes. 

In line 5, there's no need to add "framework" because "Micronaut" is being used as a modifier for "capabilities." 

It isn't clear whether you'd like "job" to be treated as a proper noun; I capitalized it for consistency.

* Proofreading (#549)

Minor changes. Also, on Line 80, should that read, "The previous service **returns** a 'List<book>'"?

* Proofreading (#548)

Minor changes

* Proofreading (#547)

Minor changes.

* Proofreading (#546)

Minor corrections

* Proofreading (#545)

Several minor corrections. Also, on line 201, it looks like there may be a typo. It says, "The previous request generates a traces composed by 9 spans." "A traces" doesn't sound right to me.

Co-authored-by: Kristin Kroeger <33693760+kroegerk@users.noreply.github.com>