Azure Micronaut Data MySQL Guide (#771)
* allow apps to have no features

* micronaut-cloud-database base guide and Oracle Cloud guide

* remove slug and asciidoctor

* add ingress rule step and curl examples

* only check output dir exists once

* updated apps merge logic

retain apps that are only in the base metadata or only in guide metadata, then merge features from apps in both

* delete .gitkeep

* PR feedback from Hari

highlight that there's no free-tier MySQL
add PowerShell cli info

* Update metadata.json (#768)

* Add azure guide

* correct output

* typo

* remove Oracle Cloud from shared

* Proofreading fixes

* update publication date

* address feedback

* add github workflow guides

* remove Graeme Rocher as base author

* add localhost:8080

* Apply suggestions from code review

Co-authored-by: Kristin Kroeger <33693760+kroegerk@users.noreply.github.com>

* Proofreading (#797)

The instructions seemed to be out of order. It said the first thing the reader would do was to create a resource group, but then it jumped backwards and said the first thing to do was configure environment variables. I reordered the sections so the environment variables come first, to make it more clear.

Other changes are general corrections and recommendations for clarity and consistency.

* revert sections

* change azureregion copy

Co-authored-by: Burt Beckwith <burt@burtbeckwith.com>
Co-authored-by: Sergio del Amo <sergio.delamo@softamo.com>
Co-authored-by: Kristin Kroeger <33693760+kroegerk@users.noreply.github.com>