revert the guide to use H2 (#969)
I prefer to avoid the usage in H2 in guides. I would like for use not to produce any documentation which uses H2.

I have seen people using H2 for test and PostgresSQL in Micronaut even if we have made it difficutl. I don't want us to encourage that behaviour.

But if you feel strongly the metrics guide is better using H2. This reverts my change to use MySQL.