Fix maven mongo guides
There was an issue with Mongo under Test Resources v1.1.0.

There is a new Gradle plugin which uses this new version, however we don't yet
have it in the Maven Plugin.

This fix specifies the mongo image to use, so that the tests pass under Maven
as well as Gradle.

This can be removed once we have a new Maven plugin version with a > 1.1.0 baseline