Fix Oracle tests.
Uses Test Resources for Kafka.
Uses TestContainers for Oracle.

This is because TestResources doesn't see the oracle-atp dependency as requiring the
Oracle Module.

Stopped using the weird database image from gitlab, as I don't know what it is, and the
password was wrong

Also includes the test-resources shutdown addition to test script