bug: Fix CLI command for multi-app guides
When there are multiple apps in a guide, the format for @cli-command@ should be @app-name:cli-command@

Previously, if this was missed, we outputted a blank command into the guide.

This change catches this (easy to make) bug, and throws an exception at build time like so:

```
Execution failed for task ':micronautOracleCloudStreamingGenerateDocs'.
> No CLI command found for app: default -- should be one of @chess-game:cli-command@, @chess-listener:cli-command@
```

The regular expression previously also didn't allow hyphens in the app names, this has also been
rectified