Add an await when counting entities for data-hibernate-reactive
The data-hibernate-reactive guide is flakey and holds up the entire pipeline.

Running under the assumption that under load (and with Github Agent minimal resources), sometimes
it takes time for the database to be atomically updated, so the count of genres is not correct for
an amount of time.

This PR adds awaitility and waits up to 10 seconds for the count to be correct (I believe this was
the failing assertion when we looked into this before)