CC 4.0 License for writing and Apache 2.0 for code (#1403)
* CC 4.0 License for writing and Apache 2.0 for code

Micronaut Guides was missing a License file. As discussed in the [Micronaut Foundation board of directors](https://micronaut.io/foundation/) mailing list, this PR adds a CC 4.0 License for writing and Apache 2.0 for code.

Moreover, it adds the [spotless plugin](https://github.com/diffplug/spotless) with an invisible feature to ensure the code contains the Apache 2.0 license.

I have created a separate PR, adding the license to the code in this repository.

This PR directly adds the license to the code from the starter generation.

The HTML does not show the Apache 2.0 License, but the downloadable sample application contains the license.

This PR includes the following copy at the of every guide:

> All guides are released with an Apache license 2.0 license for the code and a Creative Commons Attribution 4.0 license for the writing and media (images…​).

The site footer includes

GUIDES ARE LICENSED UNDER CC BY 4.0 FOR THE WRITING AND MEDIA AND APACHE.2.0 FOR THE CODE

* CC 4.0 License for writing and Apache 2.0 for code

Micronaut Guides was missing a License file. As discussed in the [Micronaut Foundation board of directors](https://micronaut.io/foundation/) mailing list, this PR adds a CC 4.0 License for writing and Apache 2.0 for code.

Moreover, it adds the [spotless plugin](https://github.com/diffplug/spotless) with an invisible feature to ensure the code contains the Apache 2.0 license.

I have created a separate PR, adding the license to the code in this repository.

This PR directly adds the license to the code from the starter generation.

The HTML does not show the Apache 2.0 License, but the downloadable sample application contains the license.

This PR includes the following copy at the of every guide:

> All guides are released with an Apache license 2.0 license for the code and a Creative Commons Attribution 4.0 license for the writing and media (images…​).

The site footer includes

GUIDES ARE LICENSED UNDER CC BY 4.0 FOR THE WRITING AND MEDIA AND APACHE.2.0 FOR THE CODE

* allow possibility to skip license validation per app

disable license validation for server generator

* org.gradle.jvmargs=-Xmx4G for every gradle guide

---------

Co-authored-by: Tim Yates <tim.yates@gmail.com>